requirements.txt
anthropic>=0.40.0
python-dotenv>=1.0.0
playwright>=1.40.0
pillow>=10.0.0
requests>=2.31.0
